refactor gcp client
