+++ core/helpers/pipeline_log.py
@@ -4,6 +4,7 @@ from typing import Optional
 from google.cloud import bigquery  # type: ignore
 
 import core.constants as constants
+import core.gcp_client
 import core.utils as utils
 
 
@@ -111,7 +112,7 @@ class PipelineLog:
             )
 
             # Run the query as a job and wait for it to complete.
-            utils.execute_bq_sql(query, job_config)
+            core.gcp_client.execute_bq_sql(query, job_config)
 
         except Exception as e:
             error_details = {
@@ -147,7 +148,7 @@ class PipelineLog:
                 ]
             )
 
-            exists = list(utils.execute_bq_sql(select_query, select_config))
+            exists = list(core.gcp_client.execute_bq_sql(select_query, select_config))
 
             if exists:
                 # If the record exists, update it.
@@ -171,7 +172,7 @@ class PipelineLog:
                     ]
                 )
 
-                utils.execute_bq_sql(update_query, update_config)
+                core.gcp_client.execute_bq_sql(update_query, update_config)
             else:
                 utils.logger.warning(f"No record found for site {self.site_name} on {self.delivery_date}. Update skipped.")
         except Exception as e:
@@ -208,7 +209,7 @@ class PipelineLog:
                 ]
             )
 
-            exists = list(utils.execute_bq_sql(select_query, select_config))
+            exists = list(core.gcp_client.execute_bq_sql(select_query, select_config))
 
             if exists:
                 # If the record exists and isn't already set to running, update it.
@@ -227,7 +228,7 @@ class PipelineLog:
                     ]
                 )
 
-                utils.execute_bq_sql(update_query, update_config)
+                core.gcp_client.execute_bq_sql(update_query, update_config)
             else:
                 utils.logger.warning(f"No record found for site {self.site_name} on {self.delivery_date}. Update skipped.")
         except Exception as e:
@@ -263,7 +264,7 @@ class PipelineLog:
                 ]
             )
 
-            exists = list(utils.execute_bq_sql(select_query, select_config))
+            exists = list(core.gcp_client.execute_bq_sql(select_query, select_config))
 
             if exists:
                 # If the record exists, update it.
@@ -293,7 +294,7 @@ class PipelineLog:
                     ]
                 )
                 
-                utils.execute_bq_sql(update_query, update_config)
+                core.gcp_client.execute_bq_sql(update_query, update_config)
             else:
                 utils.logger.warning(f"No record found for site {self.site_name} on {self.delivery_date}. Update skipped.")
         except Exception as e:
